// src/lib/bindings/AppEvent.generated.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventStatus } from "./EventStatus.generated";

export type AppEvent = { id: EventId, remoteId?: RemoteId, remoteCollectionId?: CollectionId, taskId?: TaskId, title: string, description?: string, startTime: string, endTime: string, rrule?: string, exdates?: Array<string>, recurringEventId?: EventId, originalStartTime?: string, status?: EventStatus, updatedAt?: string, color?: Color, bgColor?: Color, etag?: string, dirty?: boolean, isAllDay?: boolean, };